+++ eval/src/compiler/captures.rs
@@ -109,7 +109,7 @@ impl<'a> CapturesExtractor<'a> {
             Expr::Block(block) => {
                 self.eval_block_inner(block, HashMap::new())?;
             }
-            Expr::Cast { value, .. } => {
+            Expr::TypeCast { value, .. } => {
                 self.eval(value)?;
             }