+++ typing/tests/integration/annotations.rs
@@ -428,3 +428,13 @@ fn type_cast_basics() {
 
     assert_eq!(output.to_string(), "any Hash");
 }
+
+#[test]
+fn transmuting_type_via_casts() {
+    let code = "((1, 2, 3) as any) as (Num, Num)";
+    let block = F32Grammar::parse_statements(code).unwrap();
+    let mut type_env = TypeEnvironment::new();
+    let output = type_env.process_statements(&block).unwrap();
+
+    assert_eq!(output.to_string(), "(Num, Num)");
+}